104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,14 @@
 int main(void)
 {
 	int inc;
-	unsigned long n1 = 0, n2 = 1, n3;
-	
+	unsigned long long n1 = 0, n2 = 1, n3;
+
 for (inc = 0; inc < 98; inc++)
 {
 	n3 = n1 + n2;
 	printf("%lu", n3);
 	n1 = n2;
 	n2 = n3;
-	
 	if (inc == 97)
 		printf("\n");
 	else
